androidgcs: more branding changes (java libs in ground/gcs)
