+++ ground/gcs/src/libs/juavobjects/templates/uavobjectsinittemplate.java
@@ -27,9 +27,9 @@
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-package org.taulabs.uavtalk.uavobjects;
+package org.dronin.uavtalk.uavobjects;
 
-import org.taulabs.uavtalk.UAVObjectManager;
+import org.dronin.uavtalk.UAVObjectManager;
 
 public class UAVObjectsInitialize {
