+++ ground/gcs/src/libs/juavtalk/src/org/taulabs/uavtalk/UAVObjectFieldDescription.java
@@ -14,7 +14,7 @@
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-package org.taulabs.uavtalk;
+package org.dronin.uavtalk;
 
 /**
  ******************************************************************************
