+++ ground/gcs/src/libs/juavtalk/src/org/taulabs/uavtalk/ValueParser.java
@@ -1,4 +1,4 @@
-package org.taulabs.uavtalk;
+package org.dronin.uavtalk;
 /**
  * class with static functions to parse values from the UAVTalk protocol 
  *
